Refactor task history feature
[Implemented a new `framework/feature/history` module and migrated task history logic]
[Refactored Task History UI into `TaskHistoryScreen`, `TaskHistoryLayout`, and `TaskHistoryListUI` using a request-based state pattern]
[Migrated `isHistoryId` in the `Task` entity from `Int` to `Long` to support timestamp-based history tracking]
[Updated `TaskHistoryRepository`]
[Enhanced `TaskHistoryCard` to display both task creation and completion timestamps]
[Improved `TaskInfoLayout` and `InfoTaskDialogUI` to conditionally enable detail buttons based on content availability]
[Created `framework/core/model` module]
[Cleaned up legacy history-related methods in `DatabaseRepository` and `TaskHistoryViewModel`]
[Updated localized string resources and corrected date pattern formatting in `strings.xml`]

+++ framework/core/data/src/commonMain/kotlin/sqz/checklist/data/database/Entity.kt
@@ -16,7 +16,7 @@ import kotlinx.datetime.LocalDate
  * @param doingState not implemented yet.
  * @param isPin mark task is pinned or not.
  * @param isHistoryId if > 0 means the task has been checked and should only visible in history
- *        screen. Then expected will delete the task from database if `isHistoryId` > expected value.
+ *        screen. Then expected the task can delete from database only if `isHistoryId` >= 1.
  */
 @Entity(tableName = "task")
 data class Task(
@@ -25,7 +25,7 @@ data class Task(
     @ColumnInfo val createDate: LocalDate,
     @ColumnInfo val doingState: String? = null, // not implemented
     @ColumnInfo val isPin: Boolean = false,
-    @ColumnInfo val isHistoryId: Int = 0,
+    @ColumnInfo val isHistoryId: Long = 0,
 )
 
 /**
